api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
nba_api
sqlmodel
requests
psycopg2-binary
